feat(server): User Client in server.
