--- packages/server/lib/utils/request.js
@@ -1,35 +1,0 @@
-/*
-  Copyright 2020-2022 Lowdefy, Inc
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-async function request({ url, method = 'GET', body }) {
-  const res = await fetch(url, {
-    method,
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(body),
-  });
-  if (!res.ok) {
-    // TODO: check
-    const body = await res.json();
-    console.log(res);
-    console.log(body);
-    throw new Error(body.message || 'Request error');
-  }
-  return res.json();
-}
-
-export default request;
