+++ packages/server/package.json
@@ -42,15 +42,23 @@
   "dependencies": {
     "@lowdefy/actions-core": "4.0.0-alpha.8",
     "@lowdefy/api": "4.0.0-alpha.8",
-    "@lowdefy/block-utils": "4.0.0-alpha.8",
     "@lowdefy/blocks-antd": "4.0.0-alpha.8",
     "@lowdefy/blocks-basic": "4.0.0-alpha.8",
+    "@lowdefy/blocks-color-selectors": "4.0.0-alpha.8",
+    "@lowdefy/blocks-echarts": "4.0.0-alpha.8",
     "@lowdefy/blocks-loaders": "4.0.0-alpha.8",
-    "@lowdefy/engine": "4.0.0-alpha.8",
+    "@lowdefy/blocks-markdown": "4.0.0-alpha.8",
+    "@lowdefy/client": "4.0.0-alpha.8",
+    "@lowdefy/connection-axios-http": "4.0.0-alpha.8",
+    "@lowdefy/connection-mongodb": "4.0.0-alpha.8",
+    "@lowdefy/docs": "4.0.0-alpha.8",
     "@lowdefy/helpers": "4.0.0-alpha.8",
     "@lowdefy/layout": "4.0.0-alpha.8",
     "@lowdefy/node-utils": "4.0.0-alpha.8",
     "@lowdefy/operators-js": "4.0.0-alpha.8",
+    "@lowdefy/operators-mql": "4.0.0-alpha.8",
+    "@lowdefy/operators-nunjucks": "4.0.0-alpha.8",
+    "@lowdefy/operators-yaml": "4.0.0-alpha.8",
     "next": "12.0.10",
     "next-auth": "4.1.2",
     "process": "0.11.10",
@@ -60,6 +68,7 @@
   },
   "devDependencies": {
     "@lowdefy/build": "4.0.0-alpha.8",
+    "@lowdefy/docs": "4.0.0-alpha.8",
     "@next/eslint-plugin-next": "12.0.10",
     "less": "4.1.2",
     "less-loader": "10.2.0",
